requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
Django==1.11.17
django-cors-headers==2.4.0
django-rest-framework==0.1.0
djangorestframework==3.11.0
idna==2.9
pytz==2020.1
requests==2.23.0
simplejson==3.17.0
text-unidecode==1.3
urllib3==1.25.9
